Changement de classe mère en interface mère, ajout d'une classe finale pour les methodes
